chore(refactor): Done all project refactoring
