+++ Cargo.toml
@@ -7,11 +7,9 @@ edition = "2021"
 maintenance = { status = "actively-developed" }
 
 [features]
-enable-dotenv    = ["dep:dotenv"]
-enable-chunked   = []
-disable-caching  = []
-
-default = ["disable-caching"]
+default = []
+enable-caching = []
+enable-dotenv  = ["dep:dotenv"]
 
 [dependencies]
 actix = "^0.13"
@@ -23,7 +21,7 @@ actix-multipart = "^0.6"
 anyhow = "^1.0"
 async-trait = "^0.1"
 derive_builder = "^0.20"
-derive-getters = "^0.3"
+derive-getters = "^0.4"
 elasticsearch = "8.5.0-alpha.1"
 env_logger = "^0.11"
 futures = "^0.3"
@@ -32,28 +30,18 @@ serde = "^1.0"
 serde_derive = "^1.0"
 serde_json = "^1.0"
 thiserror = "^1.0"
-walkdir = "^2.4"
-
-datetime = { path = "src/crates/datetime" }
-elquery = { path = "src/crates/elquery" }
-hasher = { path = "src/crates/hasher" }
-wrappers = { path = "src/crates/wrappers" }
-
-[dependencies.reqwest]
-version = "^0.12"
-features = ["json", "multipart"]
 
 [dependencies.chrono]
 version = "^0.4"
 features = ["rustc-serialize", "serde"]
 
-[dependencies.dotenv]
-version = "^0.15"
-optional = true
-
 [dependencies.redis]
 version = "^0.25"
 features = ["aio", "tokio-comp", "connection-manager"]
+
+[dependencies.reqwest]
+version = "^0.12"
+features = ["json", "multipart"]
 
 [dependencies.tokio]
 version = "^1.36"
@@ -66,6 +54,21 @@ features = ["actix_extras", "chrono"]
 [dependencies.utoipa-swagger-ui]
 version = "^3.1"
 features = ["actix-web"]
+
+# There is optional crate that needed by selected features.
+[dependencies.dotenv]
+version = "^0.15"
+optional = true
+
+# There is local crates that imported from src/crates path.
+[dependencies.datetime]
+path = "src/crates/datetime"
+
+[dependencies.elquery]
+path = "src/crates/elquery"
+
+[dependencies.hasher]
+path = "src/crates/hasher"
 
 [[bin]]
 name = "elastic-main"
@@ -82,7 +85,3 @@ path = "src/crates/elquery/test/test.rs"
 [[test]]
 name = "hasher_tests"
 path = "src/crates/hasher/test/test.rs"
-
-[[test]]
-name = "schema_tests"
-path = "src/crates/wrappers/test/test.rs"
