--- src/crates/wrappers/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "wrappers"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-chrono = { version = "^0.4", features = ["rustc-serialize", "serde"] }
-derive_builder = "^0.12"
-serde = "^1.0"
-serde_derive = "^1.0"
-serde_json = "^1.0"
-utoipa = { version = "3", features = ["actix_extras"] }
-utoipa-swagger-ui = { version = "3", features = ["actix-web"] }
-
-datetime = { path = "../datetime" }
-hasher = { path = "../hasher" }
-
-[[test]]
-name = "schema_tests"
-path = "test/test.rs"
