+++ src/forms/cluster.rs
@@ -2,7 +2,8 @@ use derive_builder::Builder;
 use serde::{Deserialize, Serialize};
 use utoipa::{IntoParams, ToSchema};
 
-use crate::TestExample;
+// use crate::TestExample;
+use crate::forms::TestExample;
 use std::fmt::Display;
 
 #[derive(Builder, Clone, Default, Deserialize, Serialize, ToSchema)]
