+++ src/lib.rs
@@ -1,5 +1,5 @@
 mod endpoints;
 mod errors;
+pub mod forms;
 pub mod middlewares;
 pub mod services;
-pub mod swagger;
