+++ src/services/elastic/clusters/mod.rs
@@ -1,0 +1,1 @@
+mod clusters;
