+++ src/services/elastic/documents/mod.rs
@@ -1,0 +1,1 @@
+mod documents;
