+++ src/services/elastic/folders/mod.rs
@@ -1,0 +1,1 @@
+mod folders;
