+++ src/services/elastic/paginator/mod.rs
@@ -1,0 +1,1 @@
+pub mod paginator;
