+++ src/services/mod.rs
@@ -1,6 +1,8 @@
 pub mod cacher;
 pub mod elastic;
 pub mod init;
+mod notifier;
 pub mod own_engine;
 pub mod redis_cache;
 pub mod searcher;
+pub mod swagger;
